refactor(shell): extract wingman chat + chat-streaming
Split the chatRouter IIFE out of wingman/index.js into two modules:

- chat.js owns router/backend setup, backend selector UI, send flow,
  image paste, and (temporarily) the mic button; Task 5 will pull mic
  into its own module.
- chat-streaming.js encapsulates streaming message state and the
  router.onMessage / dualMode.onMessage handler bodies behind a
  createStreamingRenderer({ messagesEl }) factory, returning handlers
  that take a small ctx object for currentMode and persistChatMessage.

window.chatRouter contract unchanged (ensureConnected, disconnect,
router, dualMode, sendMessage); binding moves to index.js (mirroring
how alerts.js owns dismissAlert).
