+++ shell/js/wingman/index.js
@@ -4,13 +4,14 @@
  * Loaded from: shell/index.html as <script type="module" src="js/wingman/index.js">
  * window exports (set across the family): chatRouter, dismissAlert,
  *   openWingmanPanel, toggleWingmanPanel, updatePanelLayout.
- *   This file sets: chatRouter.
+ *   This file sets: chatRouter (from ./chat.js's initChat() return value).
  *   dismissAlert is set by ./alerts.js.
  *   openWingmanPanel, toggleWingmanPanel, updatePanelLayout are set by ./panel.js.
  */
     import { initAlerts } from './alerts.js';
     import { initScreenshot, captureScreenshotMode } from './screenshot.js';
     import { initPanel } from './panel.js';
+    import { initChat } from './chat.js';
 
     const renderer = window.__tandemRenderer;
     if (!renderer) {
@@ -555,755 +556,7 @@
     // ═══════════════════════════════════════════════
     // Chat Router — Multi-backend chat (Phase 3)
     // ═══════════════════════════════════════════════
-    const chatRouter = (() => {
-      const messagesEl = document.getElementById('chat-messages');
-      const inputEl = document.getElementById('chat-input');
-      const sendBtn = document.getElementById('chat-send-btn');
-      const typingEl = document.getElementById('typing-indicator');
-      const wsDot = document.getElementById('ws-dot');
-      const wsStatusText = document.getElementById('ws-status-text');
-
-      // Safety check
-      if (!messagesEl || !inputEl || !sendBtn || !typingEl || !wsDot || !wsStatusText) {
-        console.error('[chatRouter] Missing required DOM elements, chat will not initialize');
-        return { ensureConnected() { }, disconnect() { } };
-      }
-
-      // ── Shared helpers ──
-
-      function scrollToBottom() {
-        requestAnimationFrame(() => {
-          messagesEl.scrollTop = messagesEl.scrollHeight;
-        });
-      }
-
-      function ensureElementAtBottom(element) {
-        // Ensure the element is at the very bottom of the container
-        messagesEl.appendChild(element);
-        scrollToBottom();
-      }
-
-      function formatTime(ts) {
-        if (!ts) return '';
-        const d = new Date(typeof ts === 'number' ? (ts < 1e12 ? ts * 1000 : ts) : ts);
-        return d.toLocaleTimeString('nl-BE', { hour: '2-digit', minute: '2-digit' });
-      }
-
-      function escapeHtml(s) {
-        return s.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/\n/g, '<br>');
-      }
-
-      function appendMessage(role, text, timestamp, source, image) {
-        const sourceClass = source || 'openclaw';
-        let cls, name;
-        if (role === 'user') {
-          cls = 'user';
-          name = 'You';
-        } else if (sourceClass === 'claude') {
-          cls = 'claude';
-          name = 'Claude';
-        } else {
-          cls = 'wingman';
-          name = 'Wingman';
-        }
-        const el = document.createElement('div');
-        el.className = `chat-msg ${cls} source-${source || sourceClass}`;
-        el.innerHTML = `<div class="msg-from">${escapeHtml(name)}</div><div class="msg-text">${escapeHtml(text)}</div><div class="msg-time">${formatTime(timestamp)}</div>`;
-        // Add image if present
-        if (image) {
-          const msgText = el.querySelector('.msg-text');
-          const img = document.createElement('img');
-          img.src = `http://localhost:8765/chat/image/${image}`;
-          img.className = 'chat-msg-image';
-          img.addEventListener('click', () => window.open(img.src, '_blank'));
-          img.onerror = () => { img.style.display = 'none'; };
-          msgText.appendChild(img);
-        }
-        messagesEl.appendChild(el);
-        scrollToBottom();
-        return el;
-      }
-
-      async function persistChatMessage(from, text, image, notifyWebhook = false) {
-        if (!window.tandem?.persistChatMessage) return false;
-        try {
-          const result = await window.tandem.persistChatMessage({ from, text, image, notifyWebhook });
-          return Boolean(result?.ok);
-        } catch {
-          return false;
-        }
-      }
-
-      // ── Router setup ──
-
-      const router = new ChatRouter();
-      const openclawBackend = new OpenClawBackend();
-      const claudeBackend = new ClaudeActivityBackend();
-
-      router.register(openclawBackend);
-      router.register(claudeBackend);
-
-      // ── DualMode setup (Fase 5) ──
-      const dualMode = new DualMode(router);
-      let currentMode = 'openclaw'; // 'openclaw' | 'claude' | 'both'
-
-      // Track streaming message elements per conversation
-      let streamingMessages = new Map(); // conversationId -> { element, startTime }
-      let currentConversationId = null;
-
-      // ── Backend selector UI ──
-
-      const btnOC = document.getElementById('btn-backend-openclaw');
-      const btnCL = document.getElementById('btn-backend-claude');
-      const btnBoth = document.getElementById('btn-backend-both');
-      const dotOC = document.getElementById('dot-openclaw');
-      const dotCL = document.getElementById('dot-claude');
-      const dotBoth = document.getElementById('dot-both');
-
-      function updateBackendUI(activeId) {
-        if (btnOC) btnOC.classList.toggle('active', activeId === 'openclaw');
-        if (btnCL) btnCL.classList.toggle('active', activeId === 'claude');
-        if (btnBoth) btnBoth.classList.toggle('active', activeId === 'both');
-
-        if (activeId === 'both') {
-          const ocConn = openclawBackend.isConnected();
-          const clConn = claudeBackend.isConnected();
-          const bothConnected = ocConn && clConn;
-          const anyConnected = ocConn || clConn;
-          wsDot.style.background = anyConnected ? 'var(--success)' : 'var(--accent)';
-          if (bothConnected) {
-            wsStatusText.textContent = 'Wingman + Claude Connected';
-          } else if (ocConn) {
-            wsStatusText.textContent = 'Wingman Connected, Claude Disconnected';
-          } else if (clConn) {
-            wsStatusText.textContent = 'Wingman Disconnected, Claude Connected';
-          } else {
-            wsStatusText.textContent = 'Wingman + Claude Disconnected';
-          }
-          inputEl.placeholder = 'Message to Wingman & Claude... (@wingman/@claude for specific)';
-        } else {
-          // Single backend mode
-          const backend = router.getActive();
-          if (backend) {
-            const connected = backend.isConnected();
-            wsDot.style.background = connected ? 'var(--success)' : 'var(--accent)';
-            wsStatusText.textContent = connected ? `${backend.name} Connected` : `${backend.name} Disconnected`;
-          }
-          if (activeId === 'claude') {
-            inputEl.placeholder = 'Message to Claude...';
-          } else {
-            inputEl.placeholder = 'Message to Wingman...';
-          }
-        }
-
-        // Update typing indicator text
-        const typingText = typingEl.querySelector('span:last-child');
-        if (typingText) {
-          if (activeId === 'both') {
-            typingText.textContent = 'AI is thinking...';
-          } else if (activeId === 'claude') {
-            typingText.textContent = 'Claude is thinking...';
-          } else {
-            typingText.textContent = 'Wingman is typing...';
-          }
-        }
-      }
-
-      function switchBackend(id) {
-        // Store any locally typed user messages before clearing
-        const localRobinMessages = [];
-        for (const child of messagesEl.children) {
-          if (child.classList.contains('user') && child.dataset.localMessage === 'true') {
-            localRobinMessages.push({
-              role: 'user',
-              text: child.querySelector('.msg-text').textContent,
-              timestamp: child.querySelector('.msg-time').textContent,
-              source: 'user'
-            });
-          }
-        }
-
-        currentMode = id;
-        messagesEl.innerHTML = '';
-        streamingMessages.clear();
-        currentConversationId = null;
-
-        if (id === 'both') {
-          // In "both" mode, set router to openclaw as default but enable dual mode
-          router.setActive('openclaw');
-          dualMode.setEnabled(true);
-          // Load combined history — OpenClaw first, then Claude
-          openclawBackend.loadHistory((msgs) => {
-            for (const m of msgs) {
-              const el = appendMessage(m.role, m.text, m.timestamp, m.source, m.image);
-              el.dataset.fromHistory = 'true';
-            }
-            // Re-add local user messages
-            for (const localMsg of localRobinMessages) {
-              const el = appendMessage(localMsg.role, localMsg.text, localMsg.timestamp, localMsg.source, localMsg.image);
-              el.dataset.localMessage = 'true';
-            }
-          });
-        } else {
-          dualMode.setEnabled(false);
-          router.setActive(id);
-          if (id === 'openclaw') {
-            openclawBackend.loadHistory((msgs) => {
-              for (const m of msgs) {
-                const el = appendMessage(m.role, m.text, m.timestamp, m.source, m.image);
-                el.dataset.fromHistory = 'true';
-              }
-              // Re-add local user messages
-              for (const localMsg of localRobinMessages) {
-                const el = appendMessage(localMsg.role, localMsg.text, localMsg.timestamp, localMsg.source, localMsg.image);
-                el.dataset.localMessage = 'true';
-              }
-            });
-          } else {
-            claudeBackend._loadHistory();
-          }
-        }
-
-        updateBackendUI(id);
-
-        // Persist choice to config
-        fetch('http://localhost:8765/config', {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ general: { activeBackend: id } })
-        }).catch(() => { });
-      }
-
-      if (btnOC) btnOC.addEventListener('click', () => switchBackend('openclaw'));
-      if (btnCL) btnCL.addEventListener('click', () => switchBackend('claude'));
-      if (btnBoth) btnBoth.addEventListener('click', () => switchBackend('both'));
-
-      // ── Connection status dots ──
-
-      router.onConnectionChange((connected, backendId) => {
-        if (backendId === 'openclaw') {
-          if (dotOC) dotOC.classList.toggle('connected', connected);
-        } else if (backendId === 'claude') {
-          if (dotCL) dotCL.classList.toggle('connected', connected);
-        }
-        if (dotBoth) dotBoth.classList.toggle('connected', openclawBackend.isConnected() && claudeBackend.isConnected());
-
-        // Update status bar for current mode
-        if (currentMode === 'both') {
-          updateBackendUI('both');
-        } else if (backendId === router.getActiveId()) {
-          const backend = router.getActive();
-          const effectiveConnected = backend ? backend.isConnected() : connected;
-          wsDot.style.background = effectiveConnected ? 'var(--success)' : 'var(--accent)';
-          wsStatusText.textContent = effectiveConnected ? `${backend.name} Connected` : `${backend.name} Disconnected`;
-        }
-      });
-
-      // ── Message handling ──
-
-      // Single-backend message handler (existing behavior)
-      router.onMessage((msg, type, backendId) => {
-        if (currentMode === 'both') return; // handled by dualMode
-
-        if (type === 'historyReload') {
-          // Store any locally typed user messages before processing history
-          const localRobinMessages = [];
-          for (const child of messagesEl.children) {
-            if (child.classList.contains('user') && child.dataset.localMessage === 'true') {
-              localRobinMessages.push({
-                role: 'user',
-                text: child.querySelector('.msg-text').textContent,
-                timestamp: child.querySelector('.msg-time').textContent,
-                source: 'user'
-              });
-            }
-          }
-
-          // Finalize any active streaming messages with correct timestamp
-          for (const [convId, streamData] of streamingMessages.entries()) {
-            const timeEl = streamData.element.querySelector('.msg-time');
-            if (timeEl) timeEl.textContent = formatTime(Date.now());
-          }
-          messagesEl.innerHTML = '';
-          streamingMessages.clear();
-          currentConversationId = null;
-
-          if (Array.isArray(msg)) {
-            for (const historyMsg of msg) {
-              const el = appendMessage(historyMsg.role, historyMsg.text, historyMsg.timestamp, historyMsg.source, historyMsg.image);
-              el.dataset.fromHistory = 'true';
-            }
-          }
-
-          setTimeout(() => {
-            for (const localMsg of localRobinMessages) {
-              let alreadyExists = false;
-              for (const child of messagesEl.children) {
-                if (child.classList.contains('user') &&
-                  child.querySelector('.msg-text').textContent === localMsg.text &&
-                  child.dataset.fromHistory === 'true') {
-                  alreadyExists = true;
-                  break;
-                }
-              }
-
-              if (!alreadyExists) {
-                const el = appendMessage(localMsg.role, localMsg.text, localMsg.timestamp, localMsg.source, localMsg.image);
-                el.dataset.localMessage = 'true';
-              }
-            }
-          }, 0);
-          return;
-        }
-
-        if (msg._streaming) {
-          // Start new conversation if needed
-          if (!currentConversationId) {
-            currentConversationId = crypto.randomUUID();
-          }
-
-          let streamData = streamingMessages.get(currentConversationId);
-          if (!streamData) {
-            // Create new streaming message element - always insert at the very end
-            const element = appendMessage(msg.role, msg.text, msg.timestamp, msg.source, msg.image);
-            streamData = {
-              element,
-              startTime: Date.now(),
-              lastPosition: messagesEl.children.length - 1
-            };
-            streamingMessages.set(currentConversationId, streamData);
-          } else {
-            // Update existing streaming element content
-            streamData.element.querySelector('.msg-text').innerHTML = escapeHtml(msg.text);
-
-            // Ensure streaming element stays at the end (after any user messages sent during streaming)
-            const currentIndex = Array.from(messagesEl.children).indexOf(streamData.element);
-            const lastIndex = messagesEl.children.length - 1;
-            if (currentIndex !== lastIndex) {
-              ensureElementAtBottom(streamData.element);
-            }
-          }
-        } else {
-          const shouldPersistOpenClawFinal = backendId === 'openclaw' && msg._final && msg.text;
-          // Finalize current conversation
-          if (currentConversationId) {
-            const streamData = streamingMessages.get(currentConversationId);
-            if (streamData) {
-              // Update timestamp to show completion time
-              const timeEl = streamData.element.querySelector('.msg-time');
-              if (timeEl) timeEl.textContent = formatTime(Date.now());
-              // Move to bottom one last time
-              messagesEl.appendChild(streamData.element);
-              scrollToBottom();
-              streamingMessages.delete(currentConversationId);
-            }
-            if (backendId === 'openclaw' && msg.text) {
-              void persistChatMessage('wingman', msg.text);
-            }
-            currentConversationId = null;
-          } else if (shouldPersistOpenClawFinal) {
-            void persistChatMessage('wingman', msg.text);
-          }
-          // Only append a new element if this is NOT a final event (final reuses the streaming element)
-          if (!msg._final) {
-            appendMessage(msg.role, msg.text, msg.timestamp, msg.source, msg.image);
-            if (backendId === 'openclaw' && msg.text) {
-              void persistChatMessage('wingman', msg.text, msg.image);
-            }
-          }
-        }
-      });
-
-      // Dual-mode message handler (Fase 5) — both backends at once
-      let dualStreamingConversations = {}; // per-backend conversation tracking
-      dualMode.onMessage((msg, type, backendId) => {
-        if (type === 'historyReload') {
-          // Store any locally typed user messages before clearing
-          const localRobinMessages = [];
-          for (const child of messagesEl.children) {
-            if (child.classList.contains('user') && child.dataset.localMessage === 'true') {
-              localRobinMessages.push({
-                role: 'user',
-                text: child.querySelector('.msg-text').textContent,
-                timestamp: child.querySelector('.msg-time').textContent,
-                source: 'user'
-              });
-            }
-          }
-
-          // Clear and rebuild to avoid duplicates
-          messagesEl.innerHTML = '';
-          streamingMessages.clear();
-          dualStreamingConversations = {};
-
-          if (Array.isArray(msg)) {
-            for (const m of msg) {
-              const el = appendMessage(m.role, m.text, m.timestamp, m.source, m.image);
-              el.dataset.fromHistory = 'true';
-            }
-
-            // Re-add local user messages that aren't in history
-            for (const localMsg of localRobinMessages) {
-              const el = appendMessage(localMsg.role, localMsg.text, localMsg.timestamp, localMsg.source, localMsg.image);
-              el.dataset.localMessage = 'true';
-            }
-          }
-          return;
-        }
-
-        if (msg._streaming) {
-          // Start new conversation for this backend if needed
-          if (!dualStreamingConversations[backendId]) {
-            dualStreamingConversations[backendId] = {
-              conversationId: crypto.randomUUID(),
-              started: false
-            };
-          }
-
-          const convId = dualStreamingConversations[backendId].conversationId;
-          let streamData = streamingMessages.get(convId);
-
-          if (!streamData) {
-            // Create new streaming message element
-            const element = appendMessage(msg.role, msg.text, msg.timestamp, msg.source, msg.image);
-            streamData = {
-              element,
-              startTime: Date.now(),
-              backendId
-            };
-            streamingMessages.set(convId, streamData);
-            dualStreamingConversations[backendId].started = true;
-          } else {
-            // Update existing streaming element content
-            streamData.element.querySelector('.msg-text').innerHTML = escapeHtml(msg.text);
-
-            // Ensure streaming element stays at the end
-            const currentIndex = Array.from(messagesEl.children).indexOf(streamData.element);
-            const lastIndex = messagesEl.children.length - 1;
-            if (currentIndex !== lastIndex) {
-              ensureElementAtBottom(streamData.element);
-            }
-          }
-        } else {
-          const shouldPersistOpenClawFinal = backendId === 'openclaw' && msg._final && msg.text;
-          // Finalize conversation for this backend
-          if (dualStreamingConversations[backendId]) {
-            const convId = dualStreamingConversations[backendId].conversationId;
-            const streamData = streamingMessages.get(convId);
-            if (streamData) {
-              // Update timestamp to show completion time
-              const timeEl = streamData.element.querySelector('.msg-time');
-              if (timeEl) timeEl.textContent = formatTime(Date.now());
-              streamingMessages.delete(convId);
-            }
-            if (backendId === 'openclaw' && msg.text) {
-              void persistChatMessage('wingman', msg.text);
-            }
-            delete dualStreamingConversations[backendId];
-          } else if (shouldPersistOpenClawFinal) {
-            void persistChatMessage('wingman', msg.text);
-          }
-          if (!msg._final) {
-            appendMessage(msg.role, msg.text, msg.timestamp, msg.source, msg.image);
-            if (backendId === 'openclaw' && msg.text) {
-              void persistChatMessage('wingman', msg.text, msg.image);
-            }
-          }
-        }
-      });
-
-      router.onTyping((typing) => {
-        if (currentMode !== 'both') {
-          typingEl.classList.toggle('active', typing);
-        }
-      });
-
-      dualMode.onTyping((typing, backendId) => {
-        // In dual mode, show typing if any backend is typing
-        typingEl.classList.toggle('active', typing);
-        const typingText = typingEl.querySelector('span:last-child');
-        if (typingText && typing) {
-          const name = backendId === 'openclaw' ? 'Wingman' : 'Claude';
-          typingText.textContent = `${name} is typing...`;
-        }
-      });
-
-      router.onSwitch((id) => {
-        if (currentMode !== 'both') {
-          updateBackendUI(id);
-        }
-      });
-
-      // ── Image paste/drop support ──
-
-      let pendingImage = null; // base64 data URL
-      const imagePreviewEl = document.getElementById('chat-image-preview');
-
-      inputEl.addEventListener('paste', (e) => {
-        const items = e.clipboardData?.items;
-        if (!items) return;
-        for (const item of items) {
-          if (item.type.startsWith('image/')) {
-            e.preventDefault();
-            const blob = item.getAsFile();
-            if (!blob) return;
-            const reader = new FileReader();
-            reader.onload = () => {
-              pendingImage = reader.result;
-              showImagePreview(pendingImage);
-            };
-            reader.readAsDataURL(blob);
-            return;
-          }
-        }
-      });
-
-      inputEl.addEventListener('dragover', (e) => {
-        e.preventDefault();
-        e.dataTransfer.dropEffect = 'copy';
-      });
-
-      inputEl.addEventListener('drop', (e) => {
-        e.preventDefault();
-        const file = e.dataTransfer?.files?.[0];
-        if (file && file.type.startsWith('image/')) {
-          const reader = new FileReader();
-          reader.onload = () => {
-            pendingImage = reader.result;
-            showImagePreview(pendingImage);
-          };
-          reader.readAsDataURL(file);
-        }
-      });
-
-      function showImagePreview(dataUrl) {
-        imagePreviewEl.innerHTML = `
-          <img src="${dataUrl}" alt="Preview">
-          <button class="remove-preview" title="Remove image">✕</button>
-        `;
-        imagePreviewEl.style.display = 'block';
-        imagePreviewEl.querySelector('.remove-preview').addEventListener('click', () => {
-          clearImagePreview();
-        });
-      }
-
-      function clearImagePreview() {
-        pendingImage = null;
-        imagePreviewEl.innerHTML = '';
-        imagePreviewEl.style.display = 'none';
-      }
-
-      // ── Send message (input + button) ──
-
-      async function sendMessage() {
-        const text = inputEl.value.trim();
-
-        // Image paste: send via IPC (before text-only check)
-        if (pendingImage) {
-          const imageData = pendingImage;
-          clearImagePreview();
-          inputEl.value = '';
-          inputEl.style.height = '';
-
-          // Show local preview immediately
-          const robinMsg = appendMessage('user', text || '', Date.now(), 'user');
-          robinMsg.dataset.localMessage = 'true';
-          const msgText = robinMsg.querySelector('.msg-text');
-          const img = document.createElement('img');
-          img.src = imageData;
-          img.className = 'chat-msg-image';
-          img.addEventListener('click', () => window.open(imageData, '_blank'));
-          msgText.appendChild(img);
-
-          // Send to main process via IPC
-          if (window.tandem?.sendChatImage) {
-            window.tandem.sendChatImage(text, imageData);
-          }
-          return;
-        }
-
-        if (!text) return;
-
-        inputEl.value = '';
-        inputEl.style.height = '';
-
-        if (currentMode === 'both') {
-          // Dual mode: parse @-mentions, send to appropriate backends
-          const { target, cleanText } = DualMode.parseMention(text);
-          if (!cleanText) return;
-
-          // Check if target backend(s) connected
-          if (target === 'claude' && !claudeBackend.isConnected()) return;
-          if (target === 'openclaw' && !openclawBackend.isConnected()) return;
-          if (target === 'both' && !openclawBackend.isConnected() && !claudeBackend.isConnected()) return;
-
-          // Show user message (display original text with @-mention for clarity)
-          const robinMsg = appendMessage('user', text, Date.now(), 'user');
-          robinMsg.dataset.localMessage = 'true';
-
-          dualMode.sendMessage(text);
-        } else {
-          // Single backend mode
-          const backend = router.getActive();
-          const activeId = router.getActiveId();
-
-          // OpenClaw: send through the official gateway chat path.
-          if (activeId === 'openclaw') {
-            const robinMsg = appendMessage('user', text, Date.now(), 'user');
-            robinMsg.dataset.localMessage = 'true';
-            const sentViaGateway = await router.sendMessage(text);
-            if (sentViaGateway) {
-              void persistChatMessage('user', text);
-            } else {
-              appendMessage('assistant', '⚠️ Wingman could not reach OpenClaw.', Date.now(), 'wingman');
-            }
-          } else {
-            // For Claude, needs WebSocket connection
-            if (!backend || !backend.isConnected()) return;
-            await router.sendMessage(text);
-          }
-        }
-      }
-
-      inputEl.addEventListener('keydown', (e) => {
-        if (e.key === 'Enter' && !e.shiftKey) {
-          e.preventDefault();
-          sendMessage();
-        }
-      });
-      inputEl.addEventListener('input', () => {
-        inputEl.style.height = '';
-        inputEl.style.height = Math.min(inputEl.scrollHeight, 80) + 'px';
-      });
-      sendBtn.addEventListener('click', sendMessage);
-
-      // ── Mic button — native voice to text (Apple Speech / Whisper) ──
-      const micBtn = document.getElementById('chat-mic-btn');
-      let micRecording = false;
-      let micMediaRecorder = null;
-      let micChunks = [];
-      let micStream = null;
-
-      // Show mic button only on Linux (macOS users use system dictation Fn+Fn)
-      if (micBtn && window.tandem?.getPlatform) {
-        const platform = window.tandem.getPlatform();
-        if (platform !== 'darwin') micBtn.style.display = '';
-      }
-
-      if (micBtn) {
-        micBtn.addEventListener('click', async () => {
-          if (micRecording) {
-            // Stop recording → transcribe
-            micRecording = false;
-            micBtn.classList.remove('active');
-            micBtn.textContent = '⏳';
-            micBtn.title = 'Transcribing...';
-            if (micMediaRecorder && micMediaRecorder.state !== 'inactive') {
-              micMediaRecorder.stop();
-            }
-          } else {
-            // Start recording
-            try {
-              micStream = await navigator.mediaDevices.getUserMedia({ audio: true, video: false });
-              micChunks = [];
-              micMediaRecorder = new MediaRecorder(micStream, { mimeType: 'audio/webm' });
-
-              micMediaRecorder.ondataavailable = (e) => {
-                if (e.data.size > 0) micChunks.push(e.data);
-              };
-
-              micMediaRecorder.onstop = async () => {
-                // Stop mic stream
-                if (micStream) { micStream.getTracks().forEach(t => t.stop()); micStream = null; }
-
-                if (micChunks.length === 0) {
-                  micBtn.textContent = '🎤';
-                  micBtn.title = 'Voice input';
-                  return;
-                }
-
-                // Combine chunks → ArrayBuffer
-                const blob = new Blob(micChunks, { type: 'audio/webm' });
-                const arrayBuffer = await blob.arrayBuffer();
-
-                // Send to main process for transcription
-                try {
-                  const result = await window.tandem.transcribeAudio(arrayBuffer, 'nl-BE');
-                  if (result.ok && result.text) {
-                    const current = inputEl.value.trim();
-                    inputEl.value = current ? current + ' ' + result.text.trim() : result.text.trim();
-                    inputEl.dispatchEvent(new Event('input'));
-                    inputEl.style.height = '';
-                    inputEl.style.height = Math.min(inputEl.scrollHeight, 80) + 'px';
-                    inputEl.focus();
-                  } else {
-                    console.warn('[mic-btn] Transcription failed:', result.error);
-                  }
-                } catch (e) {
-                  console.error('[mic-btn] Transcription error:', e);
-                }
-
-                micBtn.textContent = '🎤';
-                micBtn.title = 'Voice input';
-                micMediaRecorder = null;
-                micChunks = [];
-              };
-
-              micMediaRecorder.start();
-              micRecording = true;
-              micBtn.classList.add('active');
-              micBtn.textContent = '🔴';
-              micBtn.title = 'Recording... (click to stop & transcribe)';
-
-            } catch (err) {
-              console.error('[mic-btn] Failed to start recording:', err);
-              micBtn.textContent = '🎤';
-              micBtn.title = 'Voice input';
-              if (err.name === 'NotAllowedError') {
-                alert('Microphone access denied.\n\nGo to System Settings → Privacy & Security → Microphone and enable Tandem.');
-              }
-            }
-          }
-        });
-      }
-
-      // ── Initialize ──
-
-      // Load saved backend from config, fallback to openclaw
-      router.connectAll();
-      fetch('http://localhost:8765/config')
-        .then(r => r.json())
-        .then(cfg => {
-          const saved = cfg.general && cfg.general.activeBackend;
-          if (saved === 'claude') switchBackend('claude');
-          else if (saved === 'both') switchBackend('both');
-          else switchBackend('openclaw');
-        })
-        .catch(() => switchBackend('openclaw'));
-
-      // Listen for incoming Wingman messages pushed via POST /chat API
-      if (window.tandem && window.tandem.onChatMessage) {
-        window.tandem.onChatMessage((msg) => {
-          // msg: {id, from, text, timestamp, image}
-          // Skip user messages — already shown optimistically in the UI
-          if (msg.from === 'user') return;
-          const source = msg.from; // 'wingman' or 'claude'
-          appendMessage('assistant', msg.text, msg.timestamp, source, msg.image);
-          if (messagesEl) messagesEl.scrollTop = messagesEl.scrollHeight;
-        });
-      }
-
-      return {
-        ensureConnected() { router.connectAll(); },
-        disconnect() { router.disconnectAll(); },
-        router,
-        dualMode,
-        sendMessage(text) {
-          if (currentMode === 'both') return dualMode.sendMessage(text);
-          return router.sendMessage(text);
-        }
-      };
-    })();
+    const chatRouter = initChat();
 
     // ═══════════════════════════════════════════════
     // Emergency stop + Approval System (Phase 4)
